update bazel build for multiple headers
