+++ README.md
@@ -5,6 +5,6 @@ clang, bazel
 
 ## Build
 ```
-bazel build --config=clang_config //src:shell
-./bazel-bin/src/shell
+bazel build --config=clang_config //src:shell_binary
+./bazel-bin/src/shell_binary
 ```
